requirements 1.txt
fastapi
uvicorn
PyPDF2
pillow
pymupdf
google-generativeai
langchain_community==0.2.4
langchain==0.2.3
faiss-cpu==1.8.0
sentence-transformers==3.0.1
langchain_huggingface
easyocr
streamlit-authenticator
